OpenCV/OCR/pytesseract demo/requirements.txt
olefile==0.44
Pillow==4.1.1
pytesseract==0.1.7
virtualenv==15.1.0
